# public/openapi/agent-commerce-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Veyra Agent API",
    "version": "0.1.0-beta.2",
    "description": "Machine API for Veyra on Arc Testnet — verified workflows, evidence-weighted agent reputation, Trust Gate policy decisions, counterparty selection, ERC-8183 independent evaluation, continuous trust monitoring, and Project 360.",
    "contact": {
      "name": "Veyra Team",
      "url": "https://github.com/mioku50/Veyra"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://agent-commerce-six.vercel.app",
      "description": "Arc Testnet Live Production Server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/execution/v1/mandates": {
      "post": {
        "summary": "Create Execution Mandate Challenge",
        "description": "Creates an EIP-712 execution mandate message challenge with spending caps and policy rules.",
        "operationId": "createExecutionMandate",
        "security": [],
        "responses": {
          "200": {
            "description": "Mandate EIP-712 payload ready for signing"
          }
        }
      },
      "get": {
        "summary": "List Execution Mandates by Owner",
        "description": "Returns active and historical execution mandates for a given owner wallet.",
        "operationId": "listExecutionMandates",
        "security": [],
        "parameters": [
          {
            "name": "ownerWallet",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of mandates"
          }
        }
      }
    },
    "/api/execution/v1/mandates/{mandateId}/activate": {
      "post": {
        "summary": "Activate Signed Execution Mandate",
        "description": "Validates the EIP-712 signature against the owner wallet and activates the mandate.",
        "operationId": "activateExecutionMandate",
        "security": [],
        "parameters": [
          {
            "name": "mandateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mandate activated"
          },
          "401": {
            "description": "Signature verification failed"
          }
        }
      }
    },
    "/api/execution/v1/prepare": {
      "post": {
        "summary": "Prepare Execution Intent",
        "description": "Preflights selection and policy checks, issues clearance, and prepares rail payload without committing funds.",
        "operationId": "prepareExecution",
        "security": [],
        "responses": {
          "200": {
            "description": "Prepared execution payload"
          },
          "422": {
            "description": "Preflight revalidation failed"
          }
        }
      }
    },
    "/api/execution/v1/{executionId}/execute": {
      "post": {
        "summary": "Execute Prepared Intent",
        "description": "Executes a prepared intent across ERC-8183 or x402, ingests reputation evidence, and updates state.",
        "operationId": "executePreparedIntent",
        "security": [],
        "parameters": [
          {
            "name": "executionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution result"
          },
          "422": {
            "description": "Budget or preflight rejected"
          }
        }
      }
    },
    "/api/execution/v1/autopilot": {
      "post": {
        "summary": "Run Autonomous Autopilot Execution",
        "description": "Discovers counterparty, selects winner, preflights, and executes under an active mandate.",
        "operationId": "runAutopilotExecution",
        "security": [],
        "responses": {
          "200": {
            "description": "Execution result"
          },
          "503": {
            "description": "Autopilot disabled"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "aac_live_...",
        "description": "Veyra Agent API Bearer credential token (`aac_live_...`)"
      }
    },
...
  }
}